--- module_freshman/src/main/java/com/mredrock/cyxbs/freshman/MainActivity.kt
@@ -1,21 +1,0 @@
-package com.mredrock.cyxbs.freshman
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
-import org.jetbrains.anko.find
-
-class MainActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.freshman_activity_main)
-        val rv = find<RecyclerView>(R.id.rv_fresh_item)
-        val layoutManager = LinearLayoutManager(this)
-        rv.layoutManager = layoutManager
-        val adapter = FreshAdapter()
-        rv.adapter = adapter
-
-    }
-}
