--- module_freshman/src/main/java/com/mredrock/cyxbs/freshman/bean/Placeholder.kt
@@ -1,6 +1,0 @@
-package com.mredrock.cyxbs.freshman.bean
-
-/**
- * Create by yuanbing
- * on 2019/8/1
- */
