+++ module_freshman/src/main/java/com/mredrock/cyxbs/freshman/view/activity/MainActivity.kt
@@ -2,12 +2,22 @@ package com.mredrock.cyxbs.freshman.view.activity
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import com.mredrock.cyxbs.freshman.view.adapter.FreshAdapter
 import com.mredrock.cyxbs.freshman.R
+import org.jetbrains.anko.find
 
 class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(R.layout.freshman_activity_main)
+        val rv = find<RecyclerView>(R.id.rv_fresh_item)
+        val layoutManager = LinearLayoutManager(this)
+        rv.layoutManager = layoutManager
+        val adapter = FreshAdapter()
+        rv.adapter = adapter
+
     }
 }
