+++ module_freshman/src/main/res/drawable-xhdpi/freshman_banner_top.png
