+++ module_freshman/src/main/res/drawable-xhdpi/freshman_btn_arrow.png
