+++ module_freshman/src/main/res/drawable-xhdpi/freshman_item_background.png
