+++ module_freshman/src/main/res/layout/freshman_activity_main.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainActivity">
+        tools:context=".view.activity.MainActivity">
     <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv_fresh_item"
             app:layout_constraintTop_toTopOf="parent"
