+++ module_freshman/src/main/res/layout/freshman_recycle_item_footer.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <ImageView xmlns:android="http://schemas.android.com/apk/res/android"
               android:orientation="vertical"
-              android:src="@drawable/bottom"
+              android:src="@drawable/freshman_banner_bottom"
               android:layout_width="match_parent"
               android:layout_height="wrap_content">
