+++ module_freshman/src/main/res/layout/freshman_recycle_item_header.xml
@@ -3,7 +3,7 @@
               android:orientation="vertical"
               android:layout_width="match_parent"
               android:layout_height="wrap_content"
-              android:src="@drawable/banner"
+              android:src="@drawable/freshman_banner_top"
 >
 
 </ImageView>
