+++ module_freshman/src/main/res/layout/freshman_recycle_item_text_item.xml
@@ -1,14 +1,28 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              android:orientation="vertical"
-              android:background="@drawable/item_background"
-              android:layout_width="match_parent"
-              android:layout_height="wrap_content">
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+                android:orientation="vertical"
+                android:background="@drawable/freshman_item_background"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content">
     <TextView
             android:id="@+id/tv_title"
+            android:text="默认"
+            android:layout_marginLeft="40dp"
+            android:layout_marginTop="15dp"
             android:layout_width="wrap_content" android:layout_height="wrap_content"/>
     <TextView
             android:id="@+id/tv_discript"
+            android:text="默认默认"
+            android:layout_below="@id/tv_title"
+            android:layout_marginLeft="40dp"
+            android:layout_marginTop="5dp"
+            android:layout_width="wrap_content" android:layout_height="wrap_content"/>
+    <ImageView
+            android:id="@+id/btn_arrow"
+            android:layout_alignParentEnd="true"
+            android:layout_marginTop="27dp"
+            android:layout_marginRight="20dp"
+            android:src="@drawable/freshman_btn_arrow"
             android:layout_width="wrap_content" android:layout_height="wrap_content"/>
 
-</LinearLayout>
+</RelativeLayout>
